Organize all tranquil cards under a top-level package
