+++ src/main/java/fruitymod/TranquilMod.java
@@ -10,11 +10,11 @@ import com.megacrit.cardcrawl.localization.CardStrings;
 import com.megacrit.cardcrawl.localization.RelicStrings;
 import com.megacrit.cardcrawl.rooms.AbstractRoom;
 import com.megacrit.cardcrawl.unlock.UnlockTracker;
-import fruitymod.cards.Defend_Tan;
-import fruitymod.cards.FlurryOfBlows;
-import fruitymod.cards.Strike_Tan;
-import fruitymod.cards.tranquil.Blockpocalypse;
-import fruitymod.cards.tranquil.Tranquil_FlyingKick;
+import fruitymod.tranquil.cards.Defend_Tan;
+import fruitymod.tranquil.cards.FlurryOfBlows;
+import fruitymod.tranquil.cards.Strike_Tan;
+import fruitymod.tranquil.cards.Blockpocalypse;
+import fruitymod.tranquil.cards.Tranquil_FlyingKick;
 import fruitymod.characters.TheTranquil;
 import fruitymod.patches.AbstractCardEnum;
 import fruitymod.patches.TheTranquilEnum;
