+++ src/main/java/fruitymod/tranquil/cards/Defend_Tan.java
@@ -1,4 +1,4 @@
-package fruitymod.cards.tranquil;
+package fruitymod.tranquil.cards;
 
 import basemod.abstracts.CustomCard;
 import com.megacrit.cardcrawl.actions.common.GainBlockAction;
@@ -11,24 +11,20 @@ import com.megacrit.cardcrawl.monsters.AbstractMonster;
 import fruitymod.TranquilMod;
 import fruitymod.patches.AbstractCardEnum;
 
-public class Blockpocalypse extends CustomCard {
-	public static final String ID = "Blockpocalypse";
+public class Defend_Tan extends CustomCard {
+	public static final String ID = "Defend_T";
 	private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
 	public static final String NAME = cardStrings.NAME;
 	public static final String DESCRIPTION = cardStrings.DESCRIPTION;
-	private static final int COST = 0;
-	private static final int BLOCK_AMT = 1;
-	private static final int TIMES = 5;
-	private static final String IMAGE = TranquilMod.makeCardImagePath(ID);
+	private static final int COST = 1;
+	private static final int BLOCK_AMT = 5;
+	private static final int UPGRADE_PLUS_BLOCK = 3;
 
-	public Blockpocalypse() {
-		super(ID, NAME, IMAGE, COST, DESCRIPTION, CardType.SKILL,
-				AbstractCardEnum.TRANQUIL_TAN, CardRarity.RARE,
-				CardTarget.SELF);
+	public Defend_Tan() {
+		super(ID, NAME, TranquilMod.makeCardImagePath(ID), COST, DESCRIPTION, CardType.SKILL,
+				AbstractCardEnum.TRANQUIL_TAN, CardRarity.BASIC, CardTarget.SELF);
 
 		this.baseBlock = BLOCK_AMT;
-		this.magicNumber = this.baseMagicNumber = TIMES;
-		this.exhaust = true;
 	}
 
 	@Override
@@ -36,22 +32,18 @@ public class Blockpocalypse extends CustomCard {
 		return true;
 	}
 
-	public void use(AbstractPlayer player, AbstractMonster monster) {
-		for (int i = 0; i < magicNumber; i++) {
-			AbstractDungeon.actionManager.addToBottom(new GainBlockAction(player, player, this.baseBlock));
-		}
+	public void use(AbstractPlayer p, AbstractMonster m) {
+		AbstractDungeon.actionManager.addToBottom(new GainBlockAction(p, p, this.block));
 	}
 
 	public AbstractCard makeCopy() {
-		return new Blockpocalypse();
+		return new Defend_Tan();
 	}
 
 	public void upgrade() {
 		if (!this.upgraded) {
 			upgradeName();
-			this.exhaust = false;
-			this.rawDescription = cardStrings.UPGRADE_DESCRIPTION;
-			this.initializeDescription();
+			upgradeBlock(UPGRADE_PLUS_BLOCK);
 		}
 	}
 }
